requirements.txt
asgiref==3.2.7
astroid==2.4.0
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
opencv-python==4.4.0.44
packaging==20.3
pluggy==0.13.1
postgres==3.0.0
psycopg2-binary==2.8.5
psycopg2==2.8.5
py==1.8.1
PyJWT==1.7.1
pylint==2.5.0
pylint-django==2.0.15
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==5.4.1
pytest-django==3.9.0
pytz==2019.3
requests==2.23.0
scipy==1.5.3
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
urllib3==1.25.9
wcwidth==0.1.9
wrapt==1.12.1
zipp==3.1.0